requirements.txt
torch==1.8.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html
tensorflow==2.5.0 -i https://pypi.doubanio.com/simple
tensorboard==2.5.0
numpy==1.21.4
matplotlib==3.4.2
mediapy==1.1.6
opencv-python
torchdiffeq